Refactoring CLI
